requirements.txt
requests==2.31.0
google-generativeai==0.5.3
pytz==2023.3.post1
beautifulsoup4==4.12.3
requests-oauthlib==1.3.1
schedule==1.2.1